[0.3.0] Default entity swap and intel handling
Abandoned trying to extend off of default entities and instead created a parent Hackable class that is inherited.

On the one hand, I wanted to be able to cast the mod entities as their vanilla counterparts and I tried using interfaces to make up for it. But I quickly realized I would be duplicating a lot of code, and it would make it difficult to generalize functionality should this project, or another mod utilizing this one, desire adding new entities.

This commit has potential. Getting backdoor access appears to be working as desired. Going forward, we should be able to cap what intel is retrievable by each entity - which opens up possibilities now for tiering hacking skills and for spreading out functionality to sensors and nav buoys.

Intel dumps in, which seems hacky. I want to get the hacked intel to pipe into a UI, but the UI system is greek to me.